create employee roles feature module | change generated ids from Integer to Long | add endpoint to end contracts | reconfigure data seeder

+++ src/main/java/io/nology/employeecreator/Employee/Employee.java
@@ -1,15 +1,9 @@
 package io.nology.employeecreator.employee;
 
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import io.nology.employeecreator.contract.Contract;
-import io.nology.employeecreator.role.Role;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 @Setter
 @Getter
@@ -19,7 +13,7 @@ public class Employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String givenName;
